removed a lot of warnings (Mostly parameterized generics)
git-svn-id: svn://oss.beweb.dbc.dk/sync_search/opensearch/trunk@4033 51cac136-f83d-432d-bb6f-125e83a60a2d

+++ src/dk/dbc/opensearch/common/types/ComparablePair.java
@@ -110,9 +110,9 @@ public class ComparablePair< E extends Comparable< E >, V extends Comparable< V 
  
         if ( first.equals( newpair.getFirst() ) )
         {
-                return (int)second.compareTo( (V)newpair.getSecond() );
+            return second.compareTo( (V)newpair.getSecond() );
         }
-        return (int)first.compareTo( (E)newpair.getFirst() );
+        return first.compareTo( (E)newpair.getFirst() );
         
     }
 }
